Add @NodeName macro to rename properties/methods
This is most-useful to use NodeSymbol names for methods to implement
well-known interfaces, but it could also be useful anywhere you want
Swift names and Node names for APIs to diverge.

Alternatively, we could add name parameters to the existing
`@NodeProperty` and `@NodeMethod` macros. This  was more
straight-forward to implement for a macro beginner.