Improve error when calling non-function values
The dynamic call behavior makes it easy to call random fictional APIs in
Swift code on NodeValue instances. I hit this error several times when
trying to iron out NodeIterator stuff. I was mistakenly doing this:

```swift
try nodeObj.set("prop", newValue)
```

which is legal from the compiler's perspective, but attempts to call a
non-existant function `set` on `nodeObj`. The correct Swift syntax is:

```swift
try nodeObj["prop"].set(to: newValue)
```

By including the DynamicProperty.key in the error message, it's much
easier to find the Swift code responsible for the error throw in
Javascript.